autocorrect: use ReplaceSurroundingText behind a flag
This CL switches to use the new API called "ReplaceSurroundingText" that
alleviates more app compat issues. Shield the new implementation behind a flag.

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4574249
Commit-Queue: Chuong Ho <hdchuong@google.com>
Cr-Commit-Position: refs/heads/main@{#1152391}

+++ chrome/browser/ash/input_method/autocorrect_manager.cc
@@ -1045,12 +1045,18 @@ void AutocorrectManager::UndoAutocorrect() {
         surrounding_text.selection_range.start() - autocorrect_range.start();
     const uint32_t after =
         autocorrect_range.end() - surrounding_text.selection_range.end();
-    input_context->DeleteSurroundingText(before, after);
 
-    // Replace with the original text.
-    input_context->CommitText(
-        pending_autocorrect_->original_text,
-        ui::TextInputClient::InsertTextCursorBehavior::kMoveCursorAfterText);
+    if (base::FeatureList::IsEnabled(
+            features::kAutocorrectUseReplaceSurroundingText)) {
+      input_context->ReplaceSurroundingText(
+          before, after, pending_autocorrect_->original_text);
+    } else {
+      input_context->DeleteSurroundingText(before, after);
+      // Replace with the original text.
+      input_context->CommitText(
+          pending_autocorrect_->original_text,
+          ui::TextInputClient::InsertTextCursorBehavior::kMoveCursorAfterText);
+    }
   }
 
   MeasureAndLogAssistiveAutocorrectQualityBreakdown(
